+++ src/app/user/user-detail/player-detail.component.ts
@@ -1,7 +1,8 @@
-import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
-import { PlayerService } from '../player.service';
-import { Player } from '../player';
+import {Component, OnInit} from '@angular/core';
+import {ActivatedRoute, Router} from '@angular/router';
+import {PlayerService} from '../player.service';
+import {Player} from '../player';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-player-detail',
@@ -22,7 +23,36 @@ export class PlayerDetailComponent implements OnInit {
   }
 
   public delete() {
-    this.playerService.delete(this.user).subscribe(
-      () => this.router.navigate(['users']));
+
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: false,
+    });
+
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonClass:"m-1",
+      confirmButtonText: 'Delete it!',
+      cancelButtonClass:"m-1",
+      cancelButtonText: 'Cancel!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.value) {
+        this.playerService.delete(this.user).subscribe(
+          () => this.router.navigate(['users']));
+        swalWithBootstrapButtons.fire(
+          'Deleted!',
+          'The user ' + this.user.username + ' has been deleted',
+          'success'
+        )
+      }
+    })
+
   }
 }
